move to environment variables
make it deployable on heroku
