+++ .gitignore
@@ -1,2 +1,3 @@
+config/env.source
 config.json
 node_modules
