--- config.sample.json
@@ -1,8 +1,0 @@
-{
-  "clientID": "",
-  "clientSecret": "",
-  "redirectURI": "https://127.0.0.1/callback",
-  "allowedOrigins": [
-    "chrome-extension://asdfasdfasdfasdf"
-  ]
-}
