+++ config/env.source.sample
@@ -1,0 +1,4 @@
+export CLIENT_ID=
+export CLIENT_SECRET=
+export REDIRECT_URI=
+export ALLOWED_ORIGIN=
